Do not require restart when a disk is unreachable during node boot
A disk that is not able to initialize when an instance is started
will never have a handler registered, which means a user will
need to restart the node after fixing the disk;

This will also prevent showing the wrong 'upgrade is needed'
error message in that case.

When the disk is still failing; print an error each 30 minutes;
Disk reconnection will be retried each 30 seconds.