[DRAFT] Cleanup deleted objects asynchronously
This PR changes existing behavior of moving permanently deleted object/version to trash right away
with a slight modification. Object versions will be marked as pending purge and queued right away if
the Delete call succeeds with quorum at the storage layer. If DELETE operation failed due to lack of quorum,this PR facilitates a graceful rollback of the Delete by undoing the previous metadata change to mark it as pending purge. This allows leaving the object intact in case there is insufficient quorum on the Delete operation.

TODO: actual undo op.