deprecate disk tokens, instead rely on deadlines and active monitoring
disk tokens usage is not necessary anymore with the implementation
of deadlines for storage calls and active monitoring of the drive
for I/O timeouts.

Functionality wise kicking off a bad drive is still supported,
its just that we do not have to serialize I/O in the manner tokens
would do. This a burden we do not have to carry forward.