automatically generate root credentials with KMS
With this commit, MinIO generates root credentials automatically
and deterministically if:

 - No root credentials have been set.
 - A KMS (KES) is configured.
 - API access for the root credentials is disabled (lockdown mode).

Before, MinIO defaults to `minioadmin` for both access key and
secret key. Now, MinIO does generate unique root credentials
automatically on startup using the KMS.

Therefore, it uses the KMS HMAC function to generate pseudo-random
values. This values never change as long as the KMS key remains
the same, and the KMS key must continue to exist sine all IAM data
is encrypted with it.

Backwards compatibility:

This commit should not cause exisiting deployments to break. It only
changes the root credentials of deployments that have a KMS configured
(KES, not a static key) but have not set any admin credentials. Such
deployments should be rare or not exist at all.

Even if, the worst case would be to update to root credentials in mc
or other clients used to administer the cluster. Root credentials
are anyway not intended for regular S3 operations.

Signed-off-by: Andreas Auernhammer <github@aead.dev>
Co-authored-by: Shubhendu <shubhendu@minio.io>
Signed-off-by: Andreas Auernhammer <github@aead.dev>