consolidate localDrives instances for proper usage
every position of a drive is static in MinIO, we do
not need complex mechanisms to infer this value anymore.
This PR is a collection of changes that shall come in
to reduce our usage of 'format.json' and removal of
duplicated instances of validation.

This PR fixes a bug that perhaps has been long introduced,
with no visible workarounds. In any deployment if an entire
erasure set is deleted, there is no way the cluster recovers.

This required a re-arranging of various pieces related
to drives and how they are updated/mutated.