Fix quorum calculation with zero parity objects
Currently the code relies on the object parity to decide if it is a
delete marker or a regular object. In case of a delete marker, the
return quorum is half of disks in the erasure set. However this
calculation is wrong with objects with EC = 0 especially because EC is not
a one time fixed configuration.

The manifested symptom is a 503 with an EC=0 object though all data are
correct. This bug is manifested after we introduced the fast Get Object
feature that does not read all data from all disks in case of inlined
objects