Handle zero versions qualified for expiration
When objects have more versions than their ILM policy expects to retain
via NewerNoncurrentVersions but they don't qualify for expiry due to
NoncurrentDays configured in that rule. In this case,
applyNewerNoncurrentVersionsLimit method was enqueuing empty tasks which
lead to a panic (panic: runtime error: index out of range [0] with
length 0) in newerNoncurrentTask.OpHash method, which assumes the task
to contain at least one version to expire.