Add drive metrics in metrics-v3
Add following metrics:

- used_inodes
- total_inodes
- healing
- online
- reads_per_sec
- reads_kb_per_sec
- reads_await
- writes_per_sec
- writes_kb_per_sec
- writes_await
- perc_util

In order to be able to calculate the `per_sec` values, we capture the
IOStats related data in the beginning (along with the time at which they
were captured), and compare them against the current values
subsequently. This is because dividing by "time since server uptime"
doesn't work in k8s environments.