list: Fix rare listing continuation freeze
Reading the list metacache is not protected by a lock, the code retries when it fails
to read the metacache object, however it forgot to re-read the metacache object
from the drives, which is necessary, especially if the metacache object is inlined.

This commit will ensure that we always re-read the metacache object from the drives
when it is retrying.