Fix RenameData data race
RenameData could start operating on inline data after timing out and the call returning due to WithDeadline.

This could cause a buffer write to the inline data being written.

Since there is no writes in `RenameData` - and the call is canceled this doesn't present a corruption issue. But a race is a race and should be fixed.

Copy inline data to a fresh buffer.