heal: Avoid marking a bucket as done when remote drives are offline
When the node healing an erasure set suddenly disconnects from the
cluster, len(disks) in the below line will always be empty,

disks, _ := er.getOnlineDisksWithHealing(false)

and that is because getOnlineDisksWithHealing(false) does not include
healing drives, hence the local drive will not be included in the list.

When this happens, a bucket is marked as done, which is simply wrong.

Requires at least N/2 non healing drives before deciding to start to
heal.