Return listing when exceeding min disk errors
When listing, with drives returning `errFileNotFound`, `errVolumeNotFound` or `errUnformattedDisk` we could get below `minDisks` drives being left.

This would result in quorum never being reachable for any object. Therefore listing would continue, but no results would ever be produced.

Include `fnf` in mindisk check, since it is incremented on these errors. This will stop listing when minDisks are left.

Allow `opts.minDisks` to not return errVolumeNotFound or errFileNotFound and return that. That will allow for good results even if there are disks returning something else.

We switch `errUnformattedDisk` be a regular error. We should fail if we have enough of those.