Support 'If-None-Match: *' for PUT case
The code for If-None-Match works if the etag is specified, but actually
the common user case for PUT is to use '*' as the match, which will match
any etag, ie will PUT the resource only if there is currently no resource
at the location. This is used for simply making sure no resource is
overwritten, and is the usual mechanism for handling concurrent updates
on object stores. This works elsewhere, eg CLoudflare R2.

Signed-off-by: Justin Cormack <justin@specialbusservice.com>