ldap: Add user DN attributes list config param
This change uses the updated ldap library in minio/pkg (which was bumped
up to v3). New config parameter is added for LDAP configuration to
specify extra user attributes to load from the LDAP server and to store
it as additional claims for the user.

A test is added in sts_handlers.go that shows how to access the LDAP
attributes as a claim.

This is in preparation to add SSH pubkey authentication to MinIO's SFTP
integration.