heal: List and heal again for any listing error
When a fresh drive healing is finished, add more check for drive listing
errors. If any, re-list and heal again. Although this is a very rare use
case to have listPathRaw() returning nil when minDisks is set to 1, we
still need to handle all possibles use cases to avoid missing healing
any object.

Also check for HealObject result to decide of an object is healed in the
fresh disk since HealObject returns nil if an object is healed in any
disk, and not in the new fresh drive.