clarify error message for root user credential
MINIO_ACCESS_KEY and MINIO_SECRET_KEY is prompted as deprecated when set in
enviroment variables. The error message on failure of credential check still
shows "access key" and "secret key", which confuses users. This patch clarifies
error messages for these cases so that it matches the actual failure reason.

Signed-off-by: Austin Chang <austin880625@gmail.com>