Log replication errors once
Also sort the errors map for multiple sites in ascending order
of deployment ids, so that the error message generated is always
definitive order and same.
Otherwise error messages were happening in random order for different
sites and so they used to not match and treated as different error
messages and logged again (comparison with old messages used to fail).

Signed-off-by: Shubhendu Ram Tripathi <shubhendu@minio.io>