Different read quorum for tiered objects
For a non-tiered object, MinIO requires that EcM (# of data blocks) of
xl.meta agree, corresponding to the number of data blocks needed to read this object.

OTOH, tiered objects have their metadata in hot tier and data in warm
tier. The data and its integrity is offloaded to the warm tier. This
allows us to reduce the read quorum from EcM (typically > N/2, where N -
erasure stripe width) to N/2 + 1. The simple majority for metadata
ensures that there is consensus on what's the object and where it is
located.