separate lock from common grid to avoid epoll contention
epoll contention on TCP causes latency build-up when
we have high volume ingress. This PR is an attempt to
relieve this pressure.

upstream issue https://github.com/golang/go/issues/65064
seems to be a deeper problem, haven't yet tried the fix
provide in this issue but however this change without
changing the compiler helps. Of course this is a workaround.