sr: Avoid recursion when loading site replicator credentials
If the site replication is enabled and the code tries to extract jwt
claims while the site replication service account credentials is still
not loaded yet, the code will enter in an infinite loop causing in a
high cpu usage.

Another possibility of the infinite loop is having some service accounts
created by an old deployment version where the service account JWT was
signed by the root credentials, but not anymore.

This commit will remove the possibility of the infinite loop in the code
and add root credential fallback to extract claims from old service
accounts.