s3: DeleteBucket to use listing before returning bucket not empty error
Use Walk(), which is a recursive listing with versioning, to check if a
bucket has some objects before allowing to remove it. This is beneficial
because the bucket can contain multiple dangling objects in multiple
drives, in a way that DeleteBucket returns bucket not empty error though
listing does not list any object.

Also, this will prevent a bug where a bucket is deleted in a deployment
that has many erasure sets but the bucket contains one or few objects
not spread in enough erasures sets.