cmd: improve multipart decryption
This commit simplifies and optimizes the decryption of large (multipart)
objects. This commit does two things:
 - Re-write the init logic for the decryption reader
 - Reduce the number of OEK decryptions

Before, the init logic copied some SSE HTTP request headers to parse
them later on. This is simplified to parsing them right away. This
removes some fields from the decryption reader struct.

Further, the decryption reader used to decrypt the OEK using the client
provided key (SSE-C) or the KMS (SSE-S3 / SSE-KMS) for each part. This
is redunant since the OEK is the same for all parts. In particular, a
KMS call might be a network request. Now, the OEK is decrypted once when
init. the reader. This should improve latency when reading encrypted
multipart objects and reduce requests to the KMS.

Signed-off-by: Andreas Auernhammer <github@aead.dev>