site-repl: Fix ILM document replication in some cases
S3 spec does not accept an ILM XML document containing both <Filter>
and <Prefix> XML tags, even if both are empty. That is why we added
'set' field in some lifecycle structures, to decide when and when not to
show a tag, however we forgot to disallow marsheling of Filter when
'set' is set to false.

This will fix ILM document replication in a site replication
configuration in some cases.