lock: Clear a contented lock not recognized by its owner
When the owner of a lock unlocks this latter, the unlock in the local
memory always succeed, but the network peers can still miss receiving
the unlock call for any reason. The current mechanism is Refresh that
will clear locks not acclaimed by its owner but a faster way is during
(R)Lock itself, when the lock is found to be held in quorum nodes but not
in the lock owner itself.